Move `spec.ts` to `spec/index.ts`
